make splitTh reduce the thesis in sight of implicit premises
Now, that's funny. I'd never thought that, say, a case hypothesis
might reduce the thesis. The case itself is (hypothesis => thesis),
right? But we can imagine proving an equivalence (A <=> B) as a
case analysis, with the cases A and B! Then the thesis is promptly
reduced to B in the first case, and to A in the second one.

Oh, no, in the second case it is not - since we take the image of
a case section as the hypothesis' negation. So after the case A,
the thesis (A <=> B) will be reduced to (- B). And in the case B,
it will be further reduced - to false. You can write somewhat
shorter proofs if you know this -- and spend some time guessing
why you assumptions are called not motivated, if you do not.