make twins work with Ind's
given how many bugs we have from exactly this kind of work,
this commit is probably the most sure way to shoot myself
in the foot. However, the most common usage of 'twins' is
in 'reduce' and 'reduce' must be cheap, so I don't want it
to instantiate bound vars. There we compare literals with
evidences of their terms, with no quantifiers in between,
so I hope this will be safe. We'll see.