refactor export
preparation of a final proof task may take a goot bit of time
which is, first, incorrectly assigned to the prover's time counter
and, second, subtracts from the prover's time limit. So we rewrite
export to prepare the task (strictly) and return an IO contunation
whose duration will be measured by reasoner.